+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -38,15 +38,11 @@ object SpringDependencies {
 }
 
 object GraphQL {
-    const val graphSpringStrarter = "com.graphql-java:graphql-spring-boot-starter:${Versions.graphSpringStarter}"
+    const val graphSpringStarter = "com.graphql-java:graphql-spring-boot-starter:${Versions.graphSpringStarter}"
     const val graphJavaTool = "com.graphql-java:graphql-java-tools:${Versions.graphJTool}"
 }
 
 object Dgs {
     const val graphPlatformDependencies = "com.netflix.graphql.dgs:graphql-dgs-platform-dependencies:latest.release"
     const val graphSpringBootStarter = "com.netflix.graphql.dgs:graphql-dgs-spring-boot-starter"
-}
-
-object ProjectLombok {
-    const val lombok = "org.projectlombok:lombok:${Versions.lombok}"
 }
