+++ src/jsMain/kotlin/Index.kt
@@ -1,4 +1,4 @@
-import frontend.App.App
+import frontend.app.app
 import react.dom.render
 import kotlinx.browser.document
 import react.create
@@ -6,5 +6,5 @@ import react.create
 fun main() {
     kotlinext.js.require("./static/style.css")
     val rootDiv = document.getElementById("root")!!
-    render(App.create(), rootDiv)
+    render(app.create(), rootDiv)
 }
