+++ src/jvmMain/kotlin/application/model/MetaData.kt
@@ -2,9 +2,6 @@ package application.model
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.elasticsearch.annotations.Document
-import org.springframework.data.elasticsearch.annotations.Field
-import org.springframework.data.elasticsearch.annotations.FieldType
-import java.util.*
 
 @Document(indexName = "test-values")
 data class MetaData(
